requirements.txt
openai-agents>=0.2.4
azure-identity
aiohttp
python-dotenv
azure-functions
azure-functions-durable